+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `discounts` object to Affirm Checkout call
+
 ## [2.0.4] - 2020-04-07
 
 ### Fixed